ambrosiaa - Ajuste no projeto para usar o Lombok

+++ backend/visseModas/src/main/java/br/com/capgemini/visseModas/ambrosia/repository/ClientePJRepository.java
@@ -3,6 +3,6 @@ package br.com.capgemini.visseModas.ambrosia.repository;
 import br.com.capgemini.visseModas.ambrosia.model.ClientePJ;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientePJRepository extends JpaRepository<ClientePJ, Integer> {
+public interface ClientePJRepository extends JpaRepository<ClientePJ, Long> {
 
 }
