+++ backend/visseModas/src/main/java/br/com/capgemini/visseModas/ambrosia/repository/EnderecoRepository.java
@@ -3,6 +3,6 @@ package br.com.capgemini.visseModas.ambrosia.repository;
 import br.com.capgemini.visseModas.ambrosia.model.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
 
 }
