+++ backend/visseModas/src/main/java/br/com/capgemini/visseModas/ambrosia/repository/ItensPedidoRepository.java
@@ -3,6 +3,6 @@ package br.com.capgemini.visseModas.ambrosia.repository;
 import br.com.capgemini.visseModas.ambrosia.model.ItensPedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItensPedidoRepository extends JpaRepository<ItensPedido, Integer> {
+public interface ItensPedidoRepository extends JpaRepository<ItensPedido, Long> {
 
 }
