+++ backend/visseModas/src/main/java/br/com/capgemini/visseModas/ambrosia/repository/ProdutoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<ProdutoRepository, Integer> {
+public interface ProdutoRepository extends JpaRepository<ProdutoRepository, Long> {
 
 }
